+++ src/Game_Render.cpp
@@ -89,7 +89,7 @@ void Game::renderBackground(uint8_t counter) {
 /* ----------------------------------------------------------------------------
  *   Render the board.
  */
-void Game::renderBoard(int xOffset, int yOffset, uint8_t topRow) { 
+void Game::renderBoard(uint8_t xOffset, uint8_t yOffset, uint8_t topRow) { 
 
 
   // Draw stripey background ..
@@ -107,9 +107,9 @@ void Game::renderBoard(int xOffset, int yOffset, uint8_t topRow) {
 
   // Draw placed pipes ..
   
-  for (int y = 0; y < puzzle.maximum.y; y++) {
+  for (uint8_t y = 0; y < puzzle.maximum.y; y++) {
       
-    for (int x = 0; x < puzzle.maximum.x; x++) {
+    for (uint8_t x = 0; x < puzzle.maximum.x; x++) {
       
       if (isPipe(x,y)) {
         
@@ -125,9 +125,9 @@ void Game::renderBoard(int xOffset, int yOffset, uint8_t topRow) {
 
   // Draw placed nodes ..
   
-  for (int y = 0; y < puzzle.maximum.y; y++) {
+  for (uint8_t y = 0; y < puzzle.maximum.y; y++) {
       
-    for (int x = 0; x < puzzle.maximum.x; x++) {
+    for (uint8_t x = 0; x < puzzle.maximum.x; x++) {
       
       if (isNode(x, y)) {
         
@@ -142,9 +142,9 @@ void Game::renderBoard(int xOffset, int yOffset, uint8_t topRow) {
   
   // Draw grid marks ..
   
-  for (int y = 0; y <= puzzle.maximum.y; y++) {
+  for (uint8_t y = 0; y <= puzzle.maximum.y; y++) {
       
-    for (int x = 0; x <= puzzle.maximum.x; x++) { 
+    for (uint8_t x = 0; x <= puzzle.maximum.x; x++) { 
         
       PD::drawPixel((x * GRID_WIDTH) + xOffset, (y * GRID_HEIGHT) + yOffset, 9);
